requirements.txt
huggingface_hub
python-dotenv
supervision
openai-agents
fastrtc
gradio
pydantic==2.10.6